done with setting up for linux

+++ src/core/main.cpp
@@ -1,3 +1,6 @@
-int CoreInit() {
+#include "macros.hpp"
+
+double CoreInit() {
 return 3.14;
 }
+
